advanced-api-project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-filter>=24.2

